braincell_backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "braincell-backend"
version = "0.1.0"
description = "BrainCell adaptive learning backend"
requires-python = ">=3.11"
dependencies = [
	"fastapi==0.115.2",
	"uvicorn[standard]==0.30.6",
	"httpx==0.27.2",
	"pydantic==1.10.18",
	"python-dotenv==1.0.1",
	"openai==1.42.0",
	"librosa==0.10.2.post1",
	"SpeechRecognition==3.10.4",
	"python-multipart==0.0.9",
	"numpy==2.1.1",
	"scipy==1.13.1",
	"sqlalchemy==2.0.23",
	"aiosqlite==0.19.0",
	"asyncpg==0.29.0",
	"psycopg2-binary==2.9.9",
]

[project.optional-dependencies]
dev = [
	"pytest==8.3.3",
	"black==24.8.0",
	"ruff==0.6.9",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
